requirements.txt
vllm>=0.4.0,<0.5.0
ray[serve]>=2.9.0
pydantic>=2.0.0
pyyaml>=6.0
